Printing prt

Printing of prt objects combines the concise yet informative design of only showing as many columns as the terminal width allows for, introduced by tibble, with the data.table approach of showing both the first and last few rows of a table. Implementation wise, the interface is designed to mimic that of tibble printing as closely as possibly, offering the same function arguments and using the same option settings (and default values) as introduced by tibble.

# S3 method for prt
print(x, ..., n = NULL, width = NULL, n_extra = NULL)

# S3 method for prt
format(x, ..., n = NULL, width = NULL, n_extra = NULL)

trunc_dt(x, n = NULL, width = NULL, n_extra = NULL)

Arguments

x
Object to format or print.

...
Other arguments passed on to individual methods.

n
Number of rows to show. If NULL, the default, will print all rows if less than option tibble.print_max. Otherwise, will print tibble.print_min rows.

width
Width of text output to generate. This defaults to NULL, which means use getOption("tibble.width") or (if also NULL) getOption("width"); the latter displays only the columns that fit on one screen. You can also set options(tibble.width = Inf) to override this default and always print all columns.

n_extra
Number of extra columns to print abbreviated information for, if the width is too small for the entire tibble. If NULL, the default, will print information about at most tibble.max_extra_cols extra columns.

Details

While the function tibble::trunc_mat() does most of the heavy lifting for formatting tibble printing output, prt exports the function trunc_dt(), which drives analogous functionality while adding the top/bottom n row concept. This function can be used for creating print() methods for other classes which represent tabular data, given that this class implements dim(), head() and tail() (and optionally tibble::tbl_sum()) methods. For an example of this, see vignette("prt", package = "prt").

The following session options are set by tibble and are respected by prt, as well as any other package that were to call trunc_dt():

tibble.print_max: Row number threshold: Maximum number of rows printed. Set to Inf to always print all rows. Default: 20.

tibble.print_min: Number of rows printed if row number threshold is exceeded. Default: 10.

tibble.width: Output width. Default: NULL (use width option).

tibble.max_extra_cols: Number of extra columns printed in reduced form. Default: 100.

Both tibble and prt rely on pillar for formatting columns and therefore, the following options set by pillar are applicable to prt printing as well.

Package options

pillar.bold: Use bold font, e.g. for column headers? This currently defaults to FALSE, because many terminal fonts have poor support for bold fonts.

pillar.subtle: Use subtle style, e.g. for row numbers and data types? Default: TRUE.

pillar.subtle_num: Use subtle style for insignificant digits? Default: FALSE, is also affected by the pillar.subtle option.

pillar.neg: Highlight negative numbers? Default: TRUE.

pillar.sigfig: The number of significant digits that will be printed and highlighted, default: 3. Set the pillar.subtle option to FALSE to turn off highlighting of significant digits.

pillar.min_title_chars: The minimum number of characters for the column title, default: 15. Column titles may be truncated up to that width to save horizontal space. Set to Inf to turn off truncation of column titles.

pillar.min_chars: The minimum number of characters wide to display character columns, default: 0. Character columns may be truncated up to that width to save horizontal space. Set to Inf to turn off truncation of character columns.

Examples

cars <- as_prt(mtcars)

print(cars)

#> # A prt:        32 × 11
#> # Partitioning: [32] rows
#>      mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1   21       6 160     110  3.9   2.62  16.5     0     1     4     4
#> 2   21       6 160     110  3.9   2.88  17.0     0     1     4     4
#> 3   22.8     4 108      93  3.85  2.32  18.6     1     1     4     1
#> 4   21.4     6 258     110  3.08  3.22  19.4     1     0     3     1
#> 5   18.7     8 360     175  3.15  3.44  17.0     0     0     3     2
#> …
#> 28  30.4     4  95.1   113  3.77  1.51  16.9     1     1     5     2
#> 29  15.8     8 351     264  4.22  3.17  14.5     0     1     5     4
#> 30  19.7     6 145     175  3.62  2.77  15.5     0     1     5     6
#> 31  15       8 301     335  3.54  3.57  14.6     0     1     5     8
#> 32  21.4     4 121     109  4.11  2.78  18.6     1     1     4     2
#> # … with 22 more rows

print(cars, n = 2)

#> # A prt:        32 × 11
#> # Partitioning: [32] rows
#>      mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1   21       6   160   110  3.9   2.62  16.5     0     1     4     4
#> 2   21       6   160   110  3.9   2.88  17.0     0     1     4     4
#> …
#> 31  15       8   301   335  3.54  3.57  14.6     0     1     5     8
#> 32  21.4     4   121   109  4.11  2.78  18.6     1     1     4     2
#> # … with 28 more rows

print(cars, width = 30)

#> # A prt:        32 × 11
#> # Partitioning: [32] rows
#>      mpg   cyl  disp    hp
#>    <dbl> <dbl> <dbl> <dbl>
#> 1   21       6 160     110
#> 2   21       6 160     110
#> 3   22.8     4 108      93
#> 4   21.4     6 258     110
#> 5   18.7     8 360     175
#> …
#> 28  30.4     4  95.1   113
#> 29  15.8     8 351     264
#> 30  19.7     6 145     175
#> 31  15       8 301     335
#> 32  21.4     4 121     109
#> # … with 22 more rows, and 7
#> #   more variables:
#> #   drat <dbl>, wt <dbl>,
#> #   qsec <dbl>, vs <dbl>,
#> #   am <dbl>, gear <dbl>,
#> #   carb <dbl>

print(cars, width = 30, n_extra = 2)

#> # A prt:        32 × 11
#> # Partitioning: [32] rows
#>      mpg   cyl  disp    hp
#>    <dbl> <dbl> <dbl> <dbl>
#> 1   21       6 160     110
#> 2   21       6 160     110
#> 3   22.8     4 108      93
#> 4   21.4     6 258     110
#> 5   18.7     8 360     175
#> …
#> 28  30.4     4  95.1   113
#> 29  15.8     8 351     264
#> 30  19.7     6 145     175
#> 31  15       8 301     335
#> 32  21.4     4 121     109
#> # … with 22 more rows, and 7
#> #   more variables:
#> #   drat <dbl>, wt <dbl>, …